+++ R/globals.R
@@ -1,6 +1,7 @@
 globalVariables(unique(c(
   # get_coord_main_resp_orga:
   "city_code_main_resp_orga", "country", "latitude", "longitude",
+  "zip_code_main_resp_orga",
   # get_canton_main_resp_orga:
   "id_canton", "short_title", "HASC_1", "geometry", "short_titleextra",
   # get_nb_cantons_influenced: